+++ KTP/KTP_Monitor/server/stats.py
@@ -153,7 +153,7 @@ def full_Stats(pupils, contest):
 
         for task in result_c:
             if result_c[task].CF in set_pupils:
-               
+                info['contest'] = all_contest
 
     info['contest'] = all_contest
     info['pupils'] = all_pupils
